requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
daphne==1.4.2
dateparser==0.6.0
Django==2.0.1
djangorestframework==3.7.7
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.5.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pymongo==3.6.0
pyOpenSSL==17.5.0
python-binance==0.6.1
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
regex==2018.1.10
requests==2.18.4
ruamel.yaml==0.15.35
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
tzlocal==1.5.1
urllib3==1.22
zope.interface==4.4.3
